+++ src/MapperGeneratorMetadataRegistryInterface.php
@@ -4,12 +4,15 @@ declare(strict_types=1);
 
 namespace AutoMapper;
 
+use AutoMapper\Transformer\CustomTransformer\CustomTransformerInterface;
 use AutoMapper\Transformer\TransformerFactoryInterface;
 
 /**
  * Registry of metadata.
  *
  * @author Joel Wurtz <jwurtz@jolicode.com>
+ *
+ * @internal
  */
 interface MapperGeneratorMetadataRegistryInterface
 {
@@ -22,6 +25,11 @@ interface MapperGeneratorMetadataRegistryInterface
      * Bind custom TransformerFactory to the AutoMapper.
      */
     public function bindTransformerFactory(TransformerFactoryInterface $transformerFactory): void;
+
+    /**
+     * Bind custom TransformerFactory to the AutoMapper.
+     */
+    public function bindCustomTransformer(CustomTransformerInterface $customTransformer): void;
 
     /**
      * Get metadata for a source and a target.