+++ src/Transformer/CustomTransformer/PrioritizedCustomTransformerInterface.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+
+namespace AutoMapper\Transformer\CustomTransformer;
+
+/**
+ * @experimental
+ */
+interface PrioritizedCustomTransformerInterface
+{
+    public function getPriority(): int;
+}