+++ tests/AutoMapperBaseTest.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace AutoMapper\Tests;
 
 use AutoMapper\AutoMapper;
+use AutoMapper\Extractor\ClassMethodToCallbackExtractor;
 use AutoMapper\Generator\Generator;
 use AutoMapper\Loader\ClassLoaderInterface;
 use AutoMapper\Loader\FileLoader;
@@ -36,6 +37,7 @@ abstract class AutoMapperBaseTest extends TestCase
         $classMetadataFactory = new ClassMetadataFactory(new AttributeLoader());
 
         $this->loader = new FileLoader(new Generator(
+            new ClassMethodToCallbackExtractor(),
             (new ParserFactory())->create(ParserFactory::PREFER_PHP7),
             new ClassDiscriminatorFromClassMetadata($classMetadataFactory),
             $allowReadOnlyTargetToPopulate