+++ tests/Fixtures/Transformer/CustomTransformer/FromSourceCustomPropertyTransformer.php
@@ -1,0 +1,21 @@
+<?php
+
+declare(strict_types=1);
+
+namespace AutoMapper\Tests\Fixtures\Transformer\CustomTransformer;
+
+use AutoMapper\Tests\Fixtures\UserDTO;
+use AutoMapper\Transformer\CustomTransformer\CustomPropertyTransformerInterface;
+
+final readonly class FromSourceCustomPropertyTransformer implements CustomPropertyTransformerInterface
+{
+    public function supports(string $source, string $target, string $propertyName): bool
+    {
+        return $source === UserDTO::class && $target === 'array' && $propertyName === 'name';
+    }
+
+    public function transform(mixed $input): mixed
+    {
+        return "{$input} set by custom property transformer";
+    }
+}