+++ tests/Fixtures/Transformer/CustomTransformer/SourceTargetCustomPropertyTransformer.php
@@ -1,0 +1,22 @@
+<?php
+
+declare(strict_types=1);
+
+namespace AutoMapper\Tests\Fixtures\Transformer\CustomTransformer;
+
+use AutoMapper\Tests\Fixtures\User;
+use AutoMapper\Tests\Fixtures\UserDTO;
+use AutoMapper\Transformer\CustomTransformer\CustomPropertyTransformerInterface;
+
+final readonly class SourceTargetCustomPropertyTransformer implements CustomPropertyTransformerInterface
+{
+    public function supports(string $source, string $target, string $propertyName): bool
+    {
+        return $source === UserDTO::class && $target === User::class && $propertyName === 'name';
+    }
+
+    public function transform(mixed $input): mixed
+    {
+        return "{$input} from custom property transformer";
+    }
+}