+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- [GH#10](https://github.com/jolicode/automapper/pull/10) Introduce custom transformers
 
 ## [8.1.0] - 2023-12-14
 ### Added